requirements.txt
torch
numpy
pandas
scikit-learn
matplotlib
pyyaml
h5py
pytest
